Add displayExpenses() Add updateBalance() Add delteTransaction()
When user enters a text, number and type a list item is shown with the input.
The balance will add up if is income or subtract if is expense and it will be reflected in the balance div.
Every item has an x, if clicked it will delete the item and the balance will either add or subtract depending whether it was income or expense.