Add cart object Add addToCart() Add changeQty()
When a user adds an item to the cart it will store that object with a new qty attribute of one. If the user adds again it will increment the qty.
When a qty button it will either increase or decrease the qty. The minus button will be disabled when is on one.